% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{aaaggregation}
\alias{aaaggregation}
\title{Scheme for amino acid aggregation}
\format{a list of length four containing one-letter name of amino acid}
\usage{
aaaggregation
}
\description{
Amino acids are grouped together in larger sets based on their
physicochemical properties important in  the recognition of signal peptide.
}
\keyword{datasets}

