% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{benchmark_dat}
\alias{benchmark_dat}
\title{Benchmark data set for signalHsmm}
\format{a list of \code{\link[seqinr]{SeqFastaAA}} objects.}
\source{
\href{http://www.uniprot.org/}{UniProt}
}
\usage{
benchmark_dat
}
\description{
Lists eukaryotic proteins added to UniProt database release 
2015_06 between 1.01.2010 and 1.06.2015 (214 proteins with signal peptide and 
214 randomly sampled proteins without signal peptide). 
Proteins longer than 80 amino acids were truncated to 80 amino acids.
}
\examples{
summary(benchmark_dat)
}
\keyword{datasets}

