% Generated by roxygen2 (4.0.1): do not edit by hand
\name{duration_viterbi}
\alias{duration_viterbi}
\title{Compute most probable path with extended Viterbi algorithm}
\usage{
duration_viterbi(aa_sample, pipar, tpmpar, od, params)
}
\arguments{
\item{aa_sample}{\code{character} vector representing single aminoacid sequence.}

\item{pipar}{Probabilities of initial state in Markov Model}

\item{tpmpar}{Matrix with transition probabilities between states}

\item{od}{Matrix of response probabilities. Eg. od[1,2] is a probability of signal 2 in state 1}

\item{params}{Matrix of probability distribution for duration.
Eg. params[10,2] is probability of duration of time 10 in state 2}
}
\value{
A list of length four:
\itemize{
\item{a vector of most probable path (\code{path},}
\item{values of probability in all intermediate points (\code{viterbi}),}
\item{matrix that for every signal and state gives the previous state in viterbi path (\code{psi}),}
\item{matrix that for every signal and state gives the duration in that state on viterbi path(\code{duration})}
}
}
\description{
Viterbi algorithm for Hidden Markov Model with duration
Valid only for special case on data
}
\note{
Currently has very restricted application to specific input All computations are on logarithms of probabilities
}

