% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_signalHsmm.R
\name{gui_signalHsmm}
\alias{gui_signalHsmm}
\title{GUI for signalHsmm}
\usage{
gui_signalHsmm()
}
\value{
null.
}
\description{
A graphical user interface for predicting presence of signal peptides.
}
\note{
Any ad-blocking software may be cause of malfunctions.
}
\seealso{
\code{\link{run_signalHsmm}}
}
