% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hsmm_pred}
\alias{hsmm_pred}
\title{hsmm_pred class}
\description{
A single prediction of \code{signal.hsmm}.
}
\details{
Always a named list of five elements
\enumerate{
\item \code{sp_probability} is a probability of signal peptide presence.
\item \code{sp_start} is a start of potential signal peptide (naively 1 aminoacid).
\item \code{sp_end} is a position of last amino acid of signal peptide.
\item \code{struc} is numeric vector representing predicted structure of input
protein.
\item \code{prot} is character vector containing input sequence of amino acids.
}
}
\seealso{
\code{\link{summary.hsmm_pred}} \code{\link{plot.hsmm_pred}}
}

