% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_protein.R
\name{is_protein}
\alias{is_protein}
\title{Protein test}
\usage{
is_protein(object)
}
\arguments{
\item{object}{\code{character} vector where each elemenents represent one amino acid.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks if an object is a protein (contains letters from one-letter amino acid code).
}
