% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.hsmm_pred}
\alias{plot.hsmm_pred}
\title{Plot single signal.hsmm prediction}
\usage{
\method{plot}{hsmm_pred}(x, add_legend = TRUE)
}
\arguments{
\item{object}{of class \code{\link{hsmm_pred}}.}

\item{if}{\code{TRUE}, legend is added to the plot.}
}
\value{
Nothing.
}
\description{
Plot objects of class \code{\link{hsmm_pred}}.
}

