% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.seq_hsmm}
\alias{predict.seq_hsmm}
\title{Predict signal.hsmm_model object}
\usage{
\method{predict}{seq_hsmm}(object, newdata, ...)
}
\arguments{
\item{object}{\code{signal.hsmm_model} object.}

\item{newdata}{unknown sequence of class \code{character} or \code{character}.
Alternatively, a \code{list} of sequences in mentioned formats.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Predict signal.hsmm_model object
}

