% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.sighsmm_model}
\alias{predict.sighsmm_model}
\title{Predict sighsmm_model object}
\usage{
\method{predict}{sighsmm_model}(object, newdata, ...)
}
\arguments{
\item{object}{\code{sighsmm_model} object.}

\item{newdata}{unknown sequence of class \code{character} or \code{character}.
Alternatively, a \code{list} of sequences in mentioned formats.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Predicts
}

