% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_txt}
\alias{read_txt}
\title{Read sequences from .txt file}
\usage{
read_txt(connection)
}
\arguments{
\item{connection}{a \code{\link{connection}} to text file.}
}
\value{
a list of sequences. Each element has class \code{\link[seqinr]{SeqFastaAA}}.
}
\description{
Read sequence data saved in text file.
}
\details{
Input file should contain one or more amino acid sequences separated by empty
rows.
}
\keyword{manip}

