% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_uniprot}
\alias{read_uniprot}
\title{Read data from UniProt database}
\usage{
read_uniprot(connection, euk)
}
\arguments{
\item{\code{\link{connection}}}{to UniProt data.}

\item{euk}{logical value if data has eukaryotic origin.}
}
\value{
a list of sequences. Each element has class \code{\link[seqinr]{SeqFastaAA}}.
Slot \code{sig} contains the range of signal peptide.
}
\description{
Read data saved in UniProt original flat text format.
}
\examples{
signal.hsmm(benchmark_dat[sample(1L:length(benchmark_dat), 3)])
}
\keyword{manip}

