% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run.signal.hsmm}
\alias{run.signal.hsmm}
\title{Predict presence of signal peptide in protein}
\usage{
run.signal.hsmm(test_data)
}
\arguments{
\item{single}{protein sequence or list of sequences. May be an object of class
\code{\link[seqinr]{SeqFastaAA}}}
}
\value{
A vector or data frame (see datails).
}
\description{
Using the hidden semi-Markov model predict presence of signal peptide in
eukaryotic proteins.
}
\details{
Function \code{signal.hsmm} returns respectively probability of presence of
signal peptide, start of signal peptide and the probable cleavage site localization.
If input consists from more than one sequence, result is a data.frame where each column
contains above values for different proteins.
}
\note{
Currently start of signal peptide is naively set as 1 amino acid.
}

