% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{signal.hsmm}
\alias{signal.hsmm}
\alias{signal.hsmm-package}
\title{signal.hsmm - prediction of signal peptides}
\description{
signal.hsmm predicts presence of signal peptides using the hidden
semi-Markov models.
}
\details{
...
}

