% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.hsmm_pred}
\alias{summary.hsmm_pred}
\title{Summarize signal.hsmm predictions}
\usage{
\method{summary}{hsmm_pred}(object, ...)
}
\arguments{
\item{object}{of class \code{\link[signal.hsmm]{hsmm_pred}}.}
}
\value{
Nothing.
}
\description{
Summarizes objects of class \code{\link[signal.hsmm]{hsmm_pred}}.
}

