% Generated by roxygen2 (4.0.2): do not edit by hand
\name{train_hsmm}
\alias{train_hsmm}
\title{Train signal.hsmm_model object}
\usage{
train_hsmm(train_data, aa_group, max.length = 32)
}
\arguments{
\item{train_data}{training data.}

\item{aa_group}{method of aggregating amino acids.}

\item{max_length}{maximum length of signal peptide.}
}
\value{
object of class \code{signal.hsmm_model}
}
\description{
Train signal.hsmm_model object
}

