% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aaaggregation}
\alias{aaaggregation}
\title{Reduced amino acid alphabet}
\format{a list of length four. Each element contains a \code{character} vector 
of amino acid names (one-letter abbreviations).}
\usage{
aaaggregation
}
\description{
Amino acids are grouped together in larger sets based on their 
physicochemical properties important in  the recognition of signal peptide.
}
\keyword{datasets}
