% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{benchmark_dat}
\alias{benchmark_dat}
\title{Benchmark data set for signal.hsmm}
\format{a list of \code{\link[seqinr]{SeqFastaAA}} objects.
Slot \code{sig} contains the range of signal peptide (if any).}
\source{
\href{http://www.uniprot.org/}{UniProt}
}
\usage{
benchmark_dat
}
\description{
This data set lists eukaryotic proteins added to UniProt database release
2014_07 between 2011 and 2014 (140 proteins with signal peptide and 280 randomly
sampled proteins without signal peptide).
All proteins were used in benchmark test comparing the performance
of signal.hsmm and other signal peptide predictors.
}
\examples{
summary(benchmark_dat)
}
\keyword{datasets}

