% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nhc.R
\name{find_nhc}
\alias{find_nhc}
\title{Localize n-, h- and c-region in signal peptide}
\usage{
find_nhc(protein, signal = NULL)
}
\arguments{
\item{protein}{a vector of amino acids or object of class 
\code{\link[seqinr]{SeqFastaAA}}.}

\item{signal}{range of signal peptide. If \code{NULL}, the attribute \code{sig}
of \code{protein} will be used.}
}
\value{
a vector of length 4 containing positions of:
\enumerate{
  \item start of n-region,
  \item start of h-region,
  \item start of c-region,
  \item cleavage site.
}
}
\description{
Finds borders between distinct regions constituting signal peptides using a heuristic
algorithm.
}
\references{
Henrik Nielsen, Anders Krogh (1998). Prediction of signal peptides
and signal anchors by a hidden Markov model. \emph{Proc. Sixth Int. Conf. on 
Intelligent Systems for Molecular Biology}.
}
