% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gui_signal.hsmm}
\alias{gui_signal.hsmm}
\title{GUI for signal.hsmm}
\usage{
gui_signal.hsmm()
}
\value{
null.
}
\description{
A graphical user interface for predicting presence of signal peptides/
}
\examples{
\donttest{
gui_signal.hsmm()
}
}
\seealso{
\code{\link{run_signal.hsmm}}
}

