% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_pred_list.R
\name{hsmm_pred_list}
\alias{hsmm_pred_list}
\title{hsmm_pred_list class}
\description{
A list of prediction(s) generated by \code{\link{run_signalHsmm}} function.
}
\details{
A named list. Each element belongs to the \code{\link{hsmm_pred}} class.
}
\seealso{
\code{\link{summary.hsmm_pred_list}}, \code{\link{pred2df}}
}
