% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hsmm_pred_list}
\alias{hsmm_pred_list}
\title{hsmm_pred_list class}
\description{
A list of prediction(s) generated by \code{\link{run_signal.hsmm}} function.
}
\details{
A named list. Each element belongs to the \code{\link{hsmm_pred}} class.
}
\seealso{
\code{\link{summary.hsmm_pred_list}}, \code{\link{pred2df}}
}

