% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_pred.R
\name{plot.hsmm_pred}
\alias{plot.hsmm_pred}
\title{Plot single signalHsmm prediction}
\usage{
\method{plot}{hsmm_pred}(x, add_legend = TRUE, only_sure = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{\link{hsmm_pred}}.}

\item{add_legend}{\code{logical}, if \code{TRUE}, legend is added to the plot.}

\item{only_sure}{\code{logical}, if \code{FALSE} does not draw signal peptide structure
when probability is smaller than 0.5.}

\item{...}{ignored.}
}
\value{
Nothing.
}
\description{
Plots objects of class \code{\link{hsmm_pred}}.
}
\keyword{hplot}
\keyword{methods}
