% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.hsmm_pred}
\alias{plot.hsmm_pred}
\title{Plot single signal.hsmm prediction}
\usage{
\method{plot}{hsmm_pred}(x, add_legend = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{\link{hsmm_pred}}.}

\item{add_legend}{logical, if \code{TRUE}, legend is added to the plot.}

\item{...}{ignored.}
}
\value{
Nothing.
}
\description{
Plot objects of class \code{\link{hsmm_pred}}.
}
\keyword{hplot}
\keyword{methods}

