% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_pred_list.R
\name{pred2df}
\alias{pred2df}
\title{Convert list of signalHsmm predictions}
\usage{
pred2df(object)
}
\arguments{
\item{object}{of class \code{\link{hsmm_pred_list}}.}
}
\value{
Data frame which columns contain respectively the probability of signal
peptide presence as well as the start and the end of predicted signal peptide.
}
\description{
Converts objects of class \code{\link{hsmm_pred_list}} to data frame.
}
\keyword{manip}
