% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sighsmm_model.R
\name{predict.sighsmm_model}
\alias{predict.sighsmm_model}
\title{Predict sighsmm_model object}
\usage{
\method{predict}{sighsmm_model}(object, newdata, ...)
}
\arguments{
\item{object}{\code{sighsmm_model} object.}

\item{newdata}{unknown sequence of class \code{character} or \code{character}.
Alternatively, a \code{list} of sequences in mentioned formats.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Predicts the presence of signal peptides using signalHsmm models.
}
\examples{
#remember to remove it
\dontrun{
pos_train_ultrahard <- read_uniprot("pos_ultrahard_data.txt", euk = TRUE)
model1 <- train_hsmm(pos_train_ultrahard, aa_group = aaaggregation)
predict(model1, benchmark_dat[1L:5])
}
}
