% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_uniprot.R
\name{read_uniprot}
\alias{read_uniprot}
\title{Read data from UniProt database}
\usage{
read_uniprot(connection, what = "signal", euk)
}
\arguments{
\item{connection}{a \code{\link{connection}} to UniProt data in text format.}

\item{what}{\code{NULL} or a single \code{character} determining which information
should be extracted. Currently can have following values: \code{"signal"},
\code{"transit"} or \code{NULL}.}

\item{euk}{logical value if data has an eukaryotic origin.}
}
\value{
a list of sequences. Each element has a class \code{\link[seqinr]{SeqFastaAA}}.
Attribute \code{sig} contains the range (start and end) of signal peptide. Attributes
\code{OS} and \code{OC} represents respectively OS and OC fields in the protein
description. Sequence with more than one cleavage site or atypical aminoacids
are removed without any notice.
}
\description{
Read data saved in UniProt original flat text format.
}
\keyword{manip}

