% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_uniprot.R
\name{read_uniprot}
\alias{read_uniprot}
\title{Read data from UniProt database}
\usage{
read_uniprot(connection, euk)
}
\arguments{
\item{connection}{a \code{\link{connection}} to UniProt data in text format.}

\item{euk}{logical value if data has an eukaryotic origin.}
}
\value{
a list of sequences. Each element has a class \code{\link[seqinr]{SeqFastaAA}}.
Slot \code{sig} contains the range (start and end) of signal peptide. Sequence with more
than one cleavage site or atypical aminoacids are removed without any notice.
}
\description{
Read data saved in UniProt original flat text format.
}
\keyword{manip}

