% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_pred.R
\name{summary.hsmm_pred}
\alias{summary.hsmm_pred}
\title{Summarize single signalHsmm prediction}
\usage{
\method{summary}{hsmm_pred}(object, only_sure = TRUE,
  double_linebreak = FALSE, ...)
}
\arguments{
\item{object}{of class \code{\link{hsmm_pred}}.}

\item{only_sure}{\code{logical}, if \code{FALSE} does not draw signal peptide structure
when probability is smaller than 0.5.}

\item{double_linebreak}{\code{logical}, if \code{TRUE} adds aestically pleasing (in .Rmd)
double linebreaks between the elements of summary.}

\item{...}{ignored}
}
\value{
Nothing.
}
\description{
Summarizes objects of class \code{\link{hsmm_pred}}.
}
\keyword{methods}
\keyword{print}
