% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_pred.R
\name{summary.hsmm_pred}
\alias{summary.hsmm_pred}
\title{Summarize single signalHsmm prediction}
\usage{
\method{summary}{hsmm_pred}(object, ...)
}
\arguments{
\item{object}{of class \code{\link{hsmm_pred}}.}

\item{...}{ignored}
}
\value{
Nothing.
}
\description{
Summarizes objects of class \code{\link{hsmm_pred}}.
}
\keyword{methods}
\keyword{print}

