% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.hsmm_pred}
\alias{summary.hsmm_pred}
\title{Summarize single signal.hsmm prediction}
\usage{
\method{summary}{hsmm_pred}(object, ...)
}
\arguments{
\item{object}{of class \code{\link{hsmm_pred}}.}

\item{...}{ignored}
}
\value{
Nothing.
}
\description{
Summarizes objects of class \code{\link{hsmm_pred}}.
}
\keyword{methods}
\keyword{print}

