% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.hsmm_pred_list}
\alias{summary.hsmm_pred_list}
\title{Summarize list of signal.hsmm predictions}
\usage{
\method{summary}{hsmm_pred_list}(object, ...)
}
\arguments{
\item{object}{of class \code{\link{hsmm_pred_list}}.}

\item{...}{ignored}
}
\value{
nothing.
}
\description{
Summarizes objects of class \code{\link{hsmm_pred_list}}.
}
\keyword{manip}
\keyword{methods}
\keyword{print}

