% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/train_hsmm.R
\name{train_hsmm}
\alias{train_hsmm}
\title{Train sighsmm_model object}
\usage{
train_hsmm(train_data, aa_group, max_length = 32)
}
\arguments{
\item{train_data}{training data.}

\item{aa_group}{method of aggregating amino acids.}

\item{max_length}{maximum length of signal peptide.}
}
\value{
object of class \code{sighsmm_model}
}
\description{
Train sighsmm_model object
}

